Adds test for graph::Manifest conversion from tracking::Manifest, and relocates Entry tests that were in the graph::Manifest test file.
